% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{GetDivisors}
\alias{GetDivisors}
\title{Get Divisors}
\usage{
GetDivisors(P)
}
\arguments{
\item{P}{a list of parameters}
}
\description{
Returns the dialect space dimentions based on matrix size and number of dialects.
}
\keyword{song-template}
