% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_BirdCharacteristics.R
\name{MakeSongCore}
\alias{MakeSongCore}
\title{Make Song Core}
\usage{
MakeSongCore(sylProb)
}
\arguments{
\item{sylProb}{the probability to know each syllable in the song core.}
}
\description{
Picks syllables based on syllProb and prevnts birds from having a syllable repertoire of 0 syllables.
}
\keyword{song-template}
